pyproject.toml
[project]
name = "coze-mcp"
version = "0.1.0"
description = "MCP server for Coze workflows"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastmcp",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
coze-mcp = "coze_mcp:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = []
